+++ task-5/1/action/IAction.java
@@ -1,4 +1,4 @@
-package controller;
+package action;
 
 public interface IAction {
     void execute();
