+++ task-5/1/action/ResolveBookRequestAction.java
@@ -1,4 +1,4 @@
-package controller;
+package action;
 
 public class ResolveBookRequestAction implements IAction {
     private final ConsoleActions consoleActions;
