+++ task-5/1/model/order/OrderStatus.java
@@ -1,4 +1,4 @@
-package model;
+package model.order;
 
 public enum OrderStatus {
     OPEN,
