+++ task-5/1/view/menu/MenuItem.java
@@ -1,6 +1,6 @@
-package view;
+package view.menu;
 
-import controller.IAction;
+import action.IAction;
 
 public class MenuItem {
     private String title;
