+++ task-5/1/view/menu/Navigator.java
@@ -1,4 +1,4 @@
-package view;
+package view.menu;
 
 import java.util.Scanner;
